Add stop-loss modification functionality
Introduced request and response DTOs for modifying stop-loss. Updated MetaTraderTrade model and order processing logic to support stop-loss modifications. Enhanced MetaApiService to handle stop-loss updates and improved order status handling.

+++ MetaTraderWorkerService/Models/MetaTraderTrade.cs
@@ -13,7 +13,7 @@ public class MetaTraderTrade
     public DateTime Time { get; set; } // Time of the trade
     public DateTime BrokerTime { get; set; } // Broker-reported time of the trade
     public DateTime UpdateTime { get; set; } // Last update time for the trade
-    public double OpenPrice { get; set; } // Opening price of the position
+    public decimal OpenPrice { get; set; } // Opening price of the position
     public decimal Volume { get; set; } // Volume of the position
     public double Swap { get; set; } // Swap fees for the position
     public double Commission { get; set; } // Commission fees for the position
